mitosheet: move utils to utils files
