+++ mitosheet/mitosheet/code_chunks/step_performers/column_steps/split_text_to_columns_code_chunk.py
@@ -7,8 +7,7 @@
 from typing import Dict, List, Optional, Union
 from mitosheet.code_chunks.code_chunk import CodeChunk
 from mitosheet.sheet_functions.types.utils import is_datetime_dtype, is_string_dtype, is_timedelta_dtype
-from mitosheet.step_performers.graph_steps.graph_utils import param_dict_to_code
-from mitosheet.transpiler.transpile_utils import column_header_list_to_transpiled_code, column_header_to_transpiled_code
+from mitosheet.transpiler.transpile_utils import column_header_list_to_transpiled_code, column_header_to_transpiled_code, param_dict_to_code
 from mitosheet.types import ColumnHeader, ColumnID
 from mitosheet.user.utils import get_pandas_version
 from mitosheet.utils import is_prev_version
