+++ mitosheet/mitosheet/transpiler/transpile_utils.py
@@ -85,5 +85,53 @@ def column_header_map_to_string(column_header_map: Dict[ColumnHeader, ColumnHead
         return result
 
 
+def param_dict_to_code(param_dict: Dict[str, Any], level: int=0, as_single_line: bool=False) -> str:
+    """
+    Takes a potentially nested params dictonary and turns it into a
+    code string that we can use in the graph generated code.
 
+    level should be 0 if we are at the highest level dict, and otherwise
+    should increment by 1 anytime we enter a new subdictonary.
+    """
+
+    # Make sure we handle as a single line properly
+    if as_single_line:
+        TAB_CONSTANT = ''
+        NEWLINE_CONSTANT = ''
+    else:
+        TAB_CONSTANT = TAB
+        NEWLINE_CONSTANT = '\n'
+
+    if level == 0:
+        code = f"{NEWLINE_CONSTANT}"
+    else:
+        code = f"dict({NEWLINE_CONSTANT}"
+
+    value_num = 0
+    for key, value in param_dict.items():
+        if isinstance(value, dict):
+            # Recurse on this nested param dictonary
+            code_chunk = f"{key} = {param_dict_to_code(value, level=level + 1)}"
+        else:
+            # We use this slighly misnamed function to make sure values get transpiled right
+            code_chunk = f"{key}={column_header_to_transpiled_code(value)}"
+        
+        # If we're not on the first value in this dict, we need to add a 
+        # command new line after the last value
+        if value_num != 0:
+            code += f", {NEWLINE_CONSTANT}"
+
+        value_num += 1
+
+        # Add spacing before the param
+        code += f"{TAB_CONSTANT * (level + 1)}"
+
+        code += f"{code_chunk}"
+
+    if level == 0:
+        code += f"{NEWLINE_CONSTANT}"
+    else:
+        # Make sure to close the dict
+        code += f"{NEWLINE_CONSTANT}{TAB_CONSTANT * (level)})"
     
+    return code
